// src/api/back/models/RulePunishmentDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Public REST api for dota2classic
 * All stuff
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface RulePunishmentDto
 */
export interface RulePunishmentDto {
    /**
     * 
     * @type {number}
     * @memberof RulePunishmentDto
     */
    id: number;
    /**
     * 
     * @type {string}
     * @memberof RulePunishmentDto
     */
    title: string;
    /**
     * 
     * @type {number}
     * @memberof RulePunishmentDto
     */
    durationHours: number;
}

export function RulePunishmentDtoFromJSON(json: any): RulePunishmentDto {
    return RulePunishmentDtoFromJSONTyped(json, false);
}

export function RulePunishmentDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): RulePunishmentDto {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'id': json['id'],
        'title': json['title'],
        'durationHours': json['durationHours'],
    };
}

export function RulePunishmentDtoToJSON(value?: RulePunishmentDto | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'id': value.id,
        'title': value.title,
        'durationHours': value.durationHours,
    };
}